server/requirements.txt
asgiref==3.5.2
certifi==2022.9.24
Django==3.2.15
django-elasticsearch-dsl==7.2.2
django-grpc==1.0.13
djangorestframework==3.14.0
elasticsearch==7.17.6
elasticsearch-dsl==7.4.0
grpcio==1.39.0
grpcio-tools==1.39.0
protobuf==3.20.3
pycodestyle==2.9.1
python-dateutil==2.8.2
pytz==2022.4
six==1.16.0
sqlparse==0.4.3
toml==0.10.2
urllib3==1.26.12
